Primite blocks to vector elements
So far only supporting dense nodes and ways for protobuf types. And ways
are not appended to vector yet.

Not sure how performant this solution is and the return type of the
function feels to abstract. Might need a step function of primitive
blocks to return one vector at a time.